Move some chromium specific tests
