+++ lib/PuppeteerSharp.Tests/ChromiumSpecificTests/PageTests.cs
@@ -1,0 +1,40 @@
+﻿using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace PuppeteerSharp.Tests.ChromiumSpecificTests
+{
+    [Collection("PuppeteerLoaderFixture collection")]
+    public class PageTests : PuppeteerPageBaseTest
+    {
+        public PageTests(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        [Fact]
+        public async Task ShouldWorkWithInterventionHeaders()
+        {
+            Server.SetRoute("/intervention", context => context.Response.WriteAsync($@"
+              <script>
+                document.write('<script src=""{TestConstants.CrossProcessHttpPrefix}/intervention.js"">' + '</scr' + 'ipt>');
+              </script>
+            "));
+            Server.SetRedirect("/intervention.js", "/redirect.js");
+
+            string interventionHeader = null;
+            Server.SetRoute("/redirect.js", context =>
+            {
+                interventionHeader = context.Request.Headers["intervention"];
+                return context.Response.WriteAsync("console.log(1);");
+            });
+
+            await Page.SetRequestInterceptionAsync(true);
+            Page.Request += async (sender, e) => await e.Request.ContinueAsync();
+
+            await Page.GoToAsync(TestConstants.ServerUrl + "/intervention");
+
+            Assert.Contains("www.chromestatus.com", interventionHeader);
+        }
+    }
+}
